# ev: Batch data for faster evaluation.